+++ cli/src/schema/content-types/shimsToContentstack.ts
@@ -1,12 +1,11 @@
 import importContentType from '#cli/cs/content-types/import.js';
 import type { ContentType } from '#cli/cs/content-types/Types.js';
-import type { SchemaFields } from '#cli/cs/Types.js';
-import isRecord from '#cli/util/isRecord.js';
 import type Ctx from '../ctx/Ctx.js';
 import isEquivalentSchema from '../isEquivalentSchema.js';
 import createProgressBar from '../lib/createProgressBar.js';
 import planMerge from '../xfer/lib/planMerge.js';
 import processPlan from '../xfer/lib/processPlan.js';
+import reduceToShim from './reduceToShim.js';
 
 // See /doc/lessons-learned/circular-references.md for the reasoning behind
 // this function.
@@ -40,37 +39,4 @@ async function noop() {
 	// no-op: content type shims are stand-ins for a full content type,
 	// and are only updated or updated when the full content type is removed
 	// or updated.
-}
-
-function reduceToShim({
-	schema: originalSchema,
-	...rest
-}: ContentType): ContentType {
-	if (!Array.isArray(originalSchema)) {
-		throw new Error('Expected schema to be an array');
-	}
-
-	const schema = [findFieldByUid(originalSchema, 'title')];
-
-	const { options } = rest;
-	if (isRecord(options) && options.is_page === true) {
-		schema.push(findFieldByUid(originalSchema, 'url'));
-	}
-
-	const tax = schema.find((x) => isRecord(x) && x.data_type === 'taxonomy');
-	if (tax) {
-		schema.push(tax);
-	}
-
-	return { ...rest, schema };
-}
-
-function findFieldByUid(schema: SchemaFields, uid: string) {
-	const match = schema.find((x) => isRecord(x) && x.uid === uid);
-
-	if (!match) {
-		throw new Error(`Expected schema to have a ${uid} field`);
-	}
-
-	return match;
 }
